move close status into socket protocol
